+++ src/app/loadout-drawer/LoadoutDrawerContents.tsx
@@ -43,7 +43,7 @@ const loadoutTypes: DimBucketType[] = [
   'Chest',
   'Leg',
   'ClassItem',
-
+  'Class',
   'Artifact',
   'Ghost',
   'Consumable',