+++ src/app/item-popup/ItemDetails.tsx
@@ -2,6 +2,7 @@ import { KillTrackerInfo } from 'app/dim-ui/KillTracker';
 import RichDestinyText from 'app/dim-ui/RichDestinyText';
 import { t } from 'app/i18next-t';
 import { storesSelector } from 'app/inventory/selectors';
+import { isTrialsPassage } from 'app/inventory/store/objectives';
 import { applySocketOverrides, useSocketOverrides } from 'app/inventory/store/override-sockets';
 import { getStore } from 'app/inventory/stores-helpers';
 import { useDefinitions } from 'app/manifest/selectors';
@@ -132,7 +133,11 @@ export default function ItemDetails({
       {defs && item.objectives && (
         <div className="item-details">
           {item.objectives.map((objective) => (
-            <Objective objective={objective} key={objective.objectiveHash} />
+            <Objective
+              objective={objective}
+              key={objective.objectiveHash}
+              isTrialsPassage={defs.isDestiny2() && isTrialsPassage(item.hash)}
+            />
           ))}
         </div>
       )}