+++ src/app/progress/PursuitItem.m.scss.d.ts
@@ -4,6 +4,7 @@ interface CssExports {
   'amount': string;
   'complete': string;
   'expired': string;
+  'flawedPassage': string;
   'fullstack': string;
   'image': string;
   'newItem': string;