+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Next
 
+* Widened and reorganized the Loadouts menu.
+  * Pull from Postmaster (and its lesser known cousin, Make room for Postmaster) are removed in favor of the button next to your Postmaster items.
+  * Randomize loadout is now at the end of the list of loadouts.
+
 ## 7.2.0 <span class="changelog-date">(2022-01-23)</span>
 
 * Weapons CSV download now includes a Zoom stat column.