+++ docs/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* Removing subclass from loadout will now enable 'Add Equipped' button
 * Widened and reorganized the Loadouts menu.
   * Pull from Postmaster (and its lesser known cousin, Make room for Postmaster) are removed in favor of the button next to your Postmaster items.
   * Randomize loadout is now at the end of the list of loadouts.